% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop6.R
\docType{data}
\name{pop6}
\alias{pop6}
\title{pop6}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1364 rows and 4 columns.
}
\usage{
data(pop6)
}
\value{
poulation
}
\description{
6th population census in China
}
\examples{
head(pop6)

}
\keyword{datasets}
