% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop7.R
\docType{data}
\name{pop7}
\alias{pop7}
\title{pop7}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1364 rows and 4 columns.
}
\usage{
data(pop7)
}
\value{
poulation
}
\description{
7th population census in China
}
\examples{
head(pop7)

}
\keyword{datasets}
