% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_age.R
\name{transform_age}
\alias{transform_age}
\title{transform_age}
\usage{
transform_age(var)
}
\arguments{
\item{var}{a age variable}
}
\value{
a new age variable
}
\description{
transform_age
}
\examples{
age <- "0岁"

transform_age(age)

path <- system.file("extdata", "pop7.xlsx", package = "zly", mustWork = TRUE)

pop7 <- readxl::read_xlsx(path)

head(pop7)

pop7 \%>\%
  dplyr::mutate(年龄 = transform_age(年龄)) \%>\%
  dplyr::sample_n(10)

}
