% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pop.R
\name{transform_pop}
\alias{transform_pop}
\title{transform_pop}
\usage{
transform_pop(path)
}
\arguments{
\item{path}{data path}
}
\value{
a tidy population data
}
\description{
transform_pop
}
\examples{
path <- system.file("extdata", "pop7raw.xlsx", package = "zly", mustWork = TRUE)
pop7_transformed <- transform_pop(path)
head(pop7_transformed)
}
