% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_region.R
\name{transform_region}
\alias{transform_region}
\title{transform_region}
\usage{
transform_region(var)
}
\arguments{
\item{var}{a region variable}
}
\value{
a new region variable
}
\description{
transform_region
}
\examples{
region <- "北京"
transform_region(region)

}
